Refactor UI Files
