+++ TheDialer/app/src/main/AndroidManifest.xml
@@ -12,6 +12,11 @@
         android:theme="@style/Theme.Dialer"
         tools:targetApi="31">
         <activity
+            android:name=".PinView"
+            android:exported="false"
+            android:label="@string/title_activity_pin_view"
+            android:theme="@style/Theme.Dialer" />
+        <activity
             android:name=".AirtimerPurchaseActivity"
             android:exported="false"
             android:label="@string/title_activity_airtimer_purchase"
