+++ TheDialer/app/src/main/res/values/strings.xml
@@ -4,4 +4,5 @@
     <string name="dial_ussd_text">Dial USSD</string>
     <string name="pick_contact_text">Pick a contact</string>
     <string name="title_activity_airtimer_purchase">AirtimerPurchaseActivity</string>
+    <string name="title_activity_pin_view">PinView</string>
 </resources>
